workers/requirements.txt
rq==1.15.1
redis==5.0.1
requests==2.31.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
